+++ client/components/Main/Mainsection/index.style.ts
@@ -44,5 +44,8 @@ export const FakeButton = styled.button`
 export const ArticlesSection = styled.div`
   background-color: ${COLORS.OFF_WHITE};
   flex: 1;
+`;
+
+export const Newsfeed = styled.div`
   overflow-y: scroll;
 `;
