+++ client/components/Main/Mainsection/index.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useRef, useState } from 'react';
 import Link from 'next/link';
 import { ArticleCard } from '../Articlecard';
-import { ArticlesSection, FakeButton, NewArticleSection, Placeholder, Wrapper } from './index.style';
+import { ArticlesSection, FakeButton, NewArticleSection, Placeholder, Wrapper, Newsfeed } from './index.style';
 import { MainTopBar } from '../../../styles/common';
 
 import Paginator, { NEXT } from '../../../utils/paginator';
@@ -30,16 +30,27 @@ export default function MainSection() {
       <MainTopBar>
         <div>홈</div>
       </MainTopBar>
-      <Link href="/write">
-        <NewArticleSection>
-          <Placeholder>무슨 생각 하세요?</Placeholder>
-          <hr />
-          <FakeButton type="button">글쓰기</FakeButton>
-        </NewArticleSection>
-      </Link>
-      <ArticlesSection>
-        {pages.map((item: any, index: number) => {
-          if (pages.length === index + 1)
+      <Newsfeed>
+        <Link href="/write">
+          <NewArticleSection>
+            <Placeholder>무슨 생각 하세요?</Placeholder>
+            <hr />
+            <FakeButton type="button">글쓰기</FakeButton>
+          </NewArticleSection>
+        </Link>
+        <ArticlesSection>
+          {pages.map((item: any, index: number) => {
+            if (pages.length === index + 1)
+              return (
+                <ArticleCard
+                  author={item.author.author}
+                  key={item.author._id}
+                  id={item.author._id}
+                  description={item.author.description}
+                  date={item.author.createdAt}
+                  ref={lastFollowElementRef}
+                />
+              );
             return (
               <ArticleCard
                 author={item.author.author}
@@ -47,20 +58,11 @@ export default function MainSection() {
                 id={item.author._id}
                 description={item.author.description}
                 date={item.author.createdAt}
-                ref={lastFollowElementRef}
               />
             );
-          return (
-            <ArticleCard
-              author={item.author.author}
-              key={item.author._id}
-              id={item.author._id}
-              description={item.author.description}
-              date={item.author.createdAt}
-            />
-          );
-        })}
-      </ArticlesSection>
+          })}
+        </ArticlesSection>
+      </Newsfeed>
     </Wrapper>
   );
 }
