+++ client/package.json
@@ -16,6 +16,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.39.7",
+    "react-loading": "^2.0.3",
     "recoil": "^0.7.6",
     "typescript": "4.8.4",
     "yup": "^0.32.11"
